feat: refactor state management to use Svelte 5 reactive stores and update component imports (#46)
---------
