+++ src/store/mobile-navbar.svelte.ts
@@ -1,0 +1,3 @@
+import { createBooleanStore } from '$/store/bool.svelte'
+
+export const isOpenMobileNavbar = createBooleanStore()
