--- src/store/mobile-navbar.ts
@@ -1,3 +1,0 @@
-import { writable } from 'svelte/store'
-
-export const isOpenMobileNavbar = writable<boolean>(false)
