+++ src/store/search.svelte.ts
@@ -1,0 +1,3 @@
+import { createBooleanStore } from '$/store/bool.svelte'
+
+export const isSearchVisible = createBooleanStore()
