--- src/store/search.ts
@@ -1,3 +1,0 @@
-import { writable } from 'svelte/store'
-
-export const isSearchVisible = writable<boolean>(false)
