--- src/store/theme.ts
@@ -1,4 +1,0 @@
-import { writable } from 'svelte/store'
-type ThemeType = 'dark' | 'light'
-
-export const theme = writable<ThemeType>('dark')
